The main file redesigned to meet class standards.
The main file has been redesinged to meet the class standards, making
sure that the operations are run perfectly from the main interface. A
big problem though remains at the level of the simulation. I have
feelings that at the moment my simulation is not generating a good
picture of the spread. Maximum work is to be done by tomorrow.